@@ -10,23 +10,21 @@
     "e2e": "ng e2e",
     "scully": "scully",
     "scully:serve": "scully serve",
-    "dev:ssr": "ng run mocolResume:serve-ssr",
-    "serve:ssr": "node dist/mocolResume/server/main.js",
-    "build:ssr": "ng build && ng run mocolResume:server",
-    "prerender": "ng run mocolResume:prerender"
+    "serve:ssr:mocolResume": "node dist/mocolResume/server/server.mjs"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^17.0.8",
+    "@angular/animations": "^18.2.13",
     "@angular/cdk": "^17.0.4",
-    "@angular/common": "^17.0.8",
-    "@angular/compiler": "^17.0.8",
-    "@angular/core": "^17.0.8",
-    "@angular/forms": "^17.0.8",
+    "@angular/common": "^18.2.13",
+    "@angular/compiler": "^18.2.13",
+    "@angular/core": "^18.2.13",
+    "@angular/forms": "^18.2.13",
     "@angular/material": "^17.0.4",
-    "@angular/platform-browser": "^17.0.8",
-    "@angular/platform-browser-dynamic": "^17.0.8",
-    "@angular/router": "^17.0.8",
+    "@angular/platform-browser": "^18.2.13",
+    "@angular/platform-browser-dynamic": "^18.2.13",
+    "@angular/router": "^18.2.13",
+    "@angular/ssr": "^18.2.20",
     "@fortawesome/angular-fontawesome": "^0.12.1",
     "@fortawesome/fontawesome-svg-core": "^6.3.0",
     "@fortawesome/free-brands-svg-icons": "^6.3.0",
@@ -41,17 +39,18 @@
     "@scullyio/scully": "2.1.41",
     "apollo-angular": "^5.0.2",
     "bootstrap": "^5.0.0",
-    "express": "^4.20.0",
+    "express": "^4.18.2",
     "graphql": "^16.8.1",
     "rxjs": "~6.6.3",
     "tslib": "^2.0.0",
-    "zone.js": "~0.14.2"
+    "zone.js": "~0.14.10"
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "^20.0.1",
-    "@angular/cli": "^17.0.9",
-    "@angular/compiler-cli": "^17.0.8",
-    "@angular/language-service": "^17.0.8",
+    "@angular/cli": "^18.2.20",
+    "@angular/compiler-cli": "^18.2.13",
+    "@angular/language-service": "^18.2.13",
+    "@types/express": "^4.17.17",
     "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "~2.0.8",
     "@types/node": "^14.11.10",
@@ -66,6 +65,6 @@
     "protractor": "~7.0.0",
     "ts-node": "~9.0.0",
     "tslint": "~6.1.3",
-    "typescript": "^5.2.2"
+    "typescript": "^5.5.4"
   }
 }
